ValeLer/requirements.txt
Flask==3.0.2
PyMySQL==1.0.2
python-dotenv==1.2.1
gunicorn==22.0.0
argon2-cffi==21.3.0